This week I went to library to work on the positioning function. Today I got two of the positioning working as they called Position 3 and Position 1. I have attached four video that I was testing in the library.
Application can't determine Position 3 or 4
This problem due to the mobile phone's Wi-Fi, because the two location are only 10 to 15 meters away. So the search result are very similar.

Position 1 working

Position 3 working, walking to Position 1

Changing from Position 3 to Position 1
There is problem that the application not update itself by press the button. 
To get the update, I need to force close the application and restart it. 
After restart it can determine I was at Position 1.